refactor: 폴더구조 재정의
